Move custom hook to separate file
