+++ lib/hooks.js
@@ -1,0 +1,25 @@
+import { auth, firestore } from '../lib/firebase';
+import { useState, useEffect } from 'react';
+import { useAuthState } from 'react-firebase-hooks/auth';
+
+export const useUserData = () => {
+  const [user] = useAuthState(auth);
+  const [username, setUsername] = useState(null);
+
+  useEffect(() => {
+    // Turn off realtime subscription
+    let unsubscribe;
+
+    if (user) {
+      const ref = firestore.collection('users').doc(user.uid);
+      unsubscribe = ref.onSnapshot((doc) => {
+        setUsername(doc.data()?.username);
+      });
+    } else {
+      setUsername(null);
+    }
+
+    return unsubscribe;
+  }, [user]);
+  return { user, username };
+};
